fix: Fix cart total price error
